+++ src/common/utils.ts
@@ -1,5 +1,3 @@
-import { decodeAddress, encodeAddress } from "@polkadot/keyring"
-import { hexToU8a, isHex } from "@polkadot/util"
 import { Buffer } from 'buffer'
 import Debug from 'debug'
 import { EntropyAccountConfig } from "../config/types"
@@ -56,16 +54,6 @@ export function getActiveOptions (options) {
 
 export function buf2hex (buffer: ArrayBuffer): string {
   return Buffer.from(buffer).toString("hex")
-}
-
-export function isValidSubstrateAddress (address: any) {
-  try {
-    encodeAddress(isHex(address) ? hexToU8a(address) : decodeAddress(address))
-
-    return true
-  } catch (error) {
-    return false
-  }
 }
 
 export function accountChoices (accounts: EntropyAccountConfig[]) {