+++ src/flows/manage-accounts/helpers/create-account.ts
@@ -3,7 +3,7 @@ import Keyring from '@entropyxyz/sdk/keys'
 import { debug } from 'src/common/utils';
 import { EntropyAccountConfig } from "src/config/types";
 
-export async function createAccount ({ name, seed, path }): Promise<EntropyAccountConfig> {
+export async function createAccount ({ name, seed, path }: { name: string, seed: string, path?: string }): Promise<EntropyAccountConfig> {
   const keyring = new Keyring({ seed, path, debug: true })
   const fullAccount = keyring.getAccount()
   // TO-DO: sdk should create account on constructor