+++ src/flows/manage-accounts/helpers/import-account.ts
@@ -35,7 +35,7 @@ export const importQuestions = [
   {
     type: 'input',
     name: 'path',
-    meesage: 'derivation path:',
+    message: 'derivation path:',
     default: 'none',
     when: ({ importKey }) => importKey
   },