+++ tests/manage-accounts.test.ts
@@ -1,7 +1,18 @@
-import { EntropyAccountConfig, EntropyConfig } from 'src/config/types'
+import { wasmGlobalsReady } from '@entropyxyz/sdk'
+// @ts-ignore
+import { isValidSubstrateAddress } from '@entropyxyz/sdk/utils'
+// @ts-ignore
+import Keyring from '@entropyxyz/sdk/keys'
+import { randomAsHex } from '@polkadot/util-crypto'
 import test from 'tape'
-import { charlieStashAddress, charlieStashSeed } from './testing-utils/constants'
+// @ts-ignore
+import { spinNetworkUp, spinNetworkDown, } from "@entropyxyz/sdk/testing"
+import { EntropyAccountConfig, EntropyConfig } from 'src/config/types'
 import { listAccounts } from 'src/flows/manage-accounts/list'
+import { createAccount } from 'src/flows/manage-accounts/helpers/create-account'
+import * as config from 'src/config'
+import { promiseRunner, sleep } from './testing-utils'
+import { charlieStashAddress, charlieStashSeed } from './testing-utils/constants'
 
 test('List Accounts', async t => {
   const account: EntropyAccountConfig = {
@@ -43,5 +54,36 @@ test('List Accounts', async t => {
     t.equal(msg, 'There are currently no accounts available, please create or import your new account using the Manage Accounts feature')
   }
 
+  t.end()
+})
+
+const networkType = 'two-nodes'
+
+let counter = 0
+test('Create Account', async t => {
+  const configPath = `/tmp/entropy-cli-${Date.now()}_${counter++}.json`
+  /* Setup */
+  const run = promiseRunner(t)
+  await run('wasm', wasmGlobalsReady())
+  await run('network up', spinNetworkUp(networkType))
+  // this gets called after all tests are run
+  t.teardown(async () => {
+    await spinNetworkDown(networkType).catch((error) =>
+      console.error('Error while spinning network down', error.message)
+    )
+  })
+  await sleep(process.env.GITHUB_WORKSPACE ? 30_000 : 5_000)
+  await run('config.init', config.init(configPath))
+  const testAccountSeed = randomAsHex(32)
+  const testAccountName = 'Test Account'
+  const newAccount = await createAccount({ name: testAccountName, seed: testAccountSeed })
+
+  const testKeyring = new Keyring({ seed: testAccountSeed, path: 'none', debug: true })
+  const { admin } = testKeyring.getAccount()
+
+  const isValidAddress = isValidSubstrateAddress(newAccount.address)
+
+  t.ok(isValidAddress, 'Valid address created')
+  t.equal(newAccount.address, admin.address, 'Generated Account matches Account created by Keyring')
   t.end()
 })